refactor: move page components into /components/pages
… and include their respective editor components
