+++ src/components/pages/AdminAttractionDetailsPage/AttractionEditor/Buttons/StatusButtons.tsx
@@ -1,55 +1,55 @@
-import { Location } from "@api/client/models/Location";
+import { AdminAttraction } from "@api/client/models/AdminAttraction";
+import Button from "@components/Button";
 import useApiClient from "@hooks/useApiClient";
-import { getAllowedStatusUpdates } from "@services/locations";
+import { getAllowedStatusUpdates } from "@services/attractions";
 import { useTranslations } from "next-intl";
 import { ReactNode, useMemo } from "react";
-import Button from "../../Button";
 
 interface Props {
-	location: Location;
+	attraction: AdminAttraction;
 	onUpdated(): void;
 }
 
-export default function StatusButtons({ location, onUpdated }: Props) {
+export default function StatusButtons({ attraction, onUpdated }: Props) {
 	const apiClient = useApiClient();
-	const t = useTranslations("Location-Details");
+	const t = useTranslations("Attraction-Details");
 	type ButtonProps = {
 		children: ReactNode;
 		onClick: () => void;
 	};
 	const buttons = useMemo<ButtonProps[]>(() => {
-		const allowedStatusUpdates = getAllowedStatusUpdates(location.status);
+		const allowedStatusUpdates = getAllowedStatusUpdates(attraction.status);
 		return [
 			allowedStatusUpdates.publish && {
 				children: t("status-update-publish"),
 				onClick: async () => {
-					await apiClient.manageCulturalData.postLocationsPublish(location.identifier);
+					await apiClient.manageCulturalData.postAttractionsPublish(attraction.identifier);
 					onUpdated();
 				},
 			},
 			allowedStatusUpdates.unpublish && {
 				children: t("status-update-unpublish"),
 				onClick: async () => {
-					await apiClient.manageCulturalData.postLocationsUnpublish(location.identifier);
+					await apiClient.manageCulturalData.postAttractionsUnpublish(attraction.identifier);
 					onUpdated();
 				},
 			},
 			allowedStatusUpdates.archive && {
 				children: t("status-update-archive"),
 				onClick: async () => {
-					await apiClient.manageCulturalData.postLocationsArchive(location.identifier);
+					await apiClient.manageCulturalData.postAttractionsArchive(attraction.identifier);
 					onUpdated();
 				},
 			},
 			allowedStatusUpdates.unarchive && {
 				children: t("status-update-unarchive"),
 				onClick: async () => {
-					await apiClient.manageCulturalData.postLocationsUnarchive(location.identifier);
+					await apiClient.manageCulturalData.postAttractionsUnarchive(attraction.identifier);
 					onUpdated();
 				},
 			},
 		].filter(Boolean) as ButtonProps[];
-	}, [apiClient, location.identifier, location.status, onUpdated, t]);
+	}, [apiClient, attraction.identifier, attraction.status, onUpdated, t]);
 	return (
 		<>
 			{buttons.map((props, index) => (
