+++ src/components/pages/AttractionsPage/index.tsx
@@ -2,11 +2,11 @@ import { Attraction } from "@api/client/models/Attraction";
 import Page from "@components/Page";
 import { getLocalizedLabel } from "@services/content";
 import { useTranslations } from "use-intl";
-import AttractionStatus from "../AttractionStatus";
-import ContentTable from "../ContentTable";
-import PageTitle from "../PageTitle";
-import Pagination, { PaginationType } from "../Pagination";
-import Spacer from "../Spacer";
+import AttractionStatus from "../../AttractionStatus";
+import ContentTable from "../../ContentTable";
+import PageTitle from "../../PageTitle";
+import Pagination, { PaginationType } from "../../Pagination";
+import Spacer from "../../Spacer";
 
 interface Props {
 	attractions: Attraction[];
