+++ src/components/pages/LocationDetailsPage/LocationEditor/Buttons/StatusButtons.tsx
@@ -1,55 +1,55 @@
-import { AdminAttraction } from "@api/client/models/AdminAttraction";
+import { Location } from "@api/client/models/Location";
+import Button from "@components/Button";
 import useApiClient from "@hooks/useApiClient";
-import { getAllowedStatusUpdates } from "@services/attractions";
+import { getAllowedStatusUpdates } from "@services/locations";
 import { useTranslations } from "next-intl";
 import { ReactNode, useMemo } from "react";
-import Button from "../../Button";
 
 interface Props {
-	attraction: AdminAttraction;
+	location: Location;
 	onUpdated(): void;
 }
 
-export default function StatusButtons({ attraction, onUpdated }: Props) {
+export default function StatusButtons({ location, onUpdated }: Props) {
 	const apiClient = useApiClient();
-	const t = useTranslations("Attraction-Details");
+	const t = useTranslations("Location-Details");
 	type ButtonProps = {
 		children: ReactNode;
 		onClick: () => void;
 	};
 	const buttons = useMemo<ButtonProps[]>(() => {
-		const allowedStatusUpdates = getAllowedStatusUpdates(attraction.status);
+		const allowedStatusUpdates = getAllowedStatusUpdates(location.status);
 		return [
 			allowedStatusUpdates.publish && {
 				children: t("status-update-publish"),
 				onClick: async () => {
-					await apiClient.manageCulturalData.postAttractionsPublish(attraction.identifier);
+					await apiClient.manageCulturalData.postLocationsPublish(location.identifier);
 					onUpdated();
 				},
 			},
 			allowedStatusUpdates.unpublish && {
 				children: t("status-update-unpublish"),
 				onClick: async () => {
-					await apiClient.manageCulturalData.postAttractionsUnpublish(attraction.identifier);
+					await apiClient.manageCulturalData.postLocationsUnpublish(location.identifier);
 					onUpdated();
 				},
 			},
 			allowedStatusUpdates.archive && {
 				children: t("status-update-archive"),
 				onClick: async () => {
-					await apiClient.manageCulturalData.postAttractionsArchive(attraction.identifier);
+					await apiClient.manageCulturalData.postLocationsArchive(location.identifier);
 					onUpdated();
 				},
 			},
 			allowedStatusUpdates.unarchive && {
 				children: t("status-update-unarchive"),
 				onClick: async () => {
-					await apiClient.manageCulturalData.postAttractionsUnarchive(attraction.identifier);
+					await apiClient.manageCulturalData.postLocationsUnarchive(location.identifier);
 					onUpdated();
 				},
 			},
 		].filter(Boolean) as ButtonProps[];
-	}, [apiClient, attraction.identifier, attraction.status, onUpdated, t]);
+	}, [apiClient, location.identifier, location.status, onUpdated, t]);
 	return (
 		<>
 			{buttons.map((props, index) => (
