+++ src/pages/admin/attractions/create.tsx
@@ -1,4 +1,4 @@
-import AdminAttractionDetailsPage from "@components/AdminAttractionDetailsPage";
+import AdminAttractionDetailsPage from "@components/pages/AdminAttractionDetailsPage";
 import { withApiClientAndPagination } from "@services/data";
 import { loadMessages } from "@services/i18n";
 import withAuth from "@services/withAuth";
