+++ src/pages/admin/attractions/index.tsx
@@ -1,5 +1,5 @@
 import { AdminAttraction } from "@api/client/models/AdminAttraction";
-import AdminAttractionsPage from "@components/AdminAttractionsPage";
+import AdminAttractionsPage from "@components/pages/AdminAttractionsPage";
 import { getPaginationProps, withApiClientAndPagination } from "@services/data";
 import withAuth from "@services/withAuth";
 import { GetServerSideProps } from "next";
