+++ src/pages/admin/organizations/[identifier].tsx
@@ -1,6 +1,6 @@
 import { Organization } from "@api/client/models/Organization";
 import { Tag } from "@common/types";
-import OrganizationDetailsPage from "@components/OrganizationDetailsPage";
+import OrganizationDetailsPage from "@components/pages/OrganizationDetailsPage";
 import { withApiClientAndPagination } from "@services/data";
 import withAuth from "@services/withAuth";
 import { GetServerSideProps } from "next";
